Struct clausy::core::formula::Formula

pub(crate) struct Formula<'a> {
    pub(crate) exprs: Vec<Expr>,
    exprs_inv: HashMap<u64, Vec<usize>>,
    pub(crate) vars: Vec<Var<'a>>,
    vars_inv: HashMap<Var<'a>, i32>,
    root_id: usize,
    var_aux_id: u32,
    new_exprs: Option<Vec<usize>>,
}

A feature-model formula.

We represent a formula by storing its syntax tree; that is, each unique sub-expression that appears in it. In canonical form, sub-expressions are uniquely stored, so no sub-expression appears twice with distinct identifiers (structural sharing). This allows for concise representation and facilitates some algorithms (e.g., Formula::to_cnf_tseitin). However, it also comes with the downside that each sub-expression has potentially many parents. Thus, owners of sub-expressions are not easily trackable (see Formula::exprs on garbage collection). Consequently, we cannot access any parents when mutating sub-expressions, only children. Due to the structural-sharing, we effectively treat the syntax tree as a directed acyclic graph. We represent this graph as an adjacency list stored in Formula::exprs. Note that due to performance reasons, structural sharing is not fully guaranteed by all algorithms (including parsers) until calling Formula::canon_visitor.

Fields
exprs: Vec<Expr>

Stores all expressions in this formula.

Serves as a fast lookup for an expression, given its identifier. Expressions are stored in the order of their creation, so new expressions are appended with Vec::push. Also, while algorithms may update expressions in-place, no expression is ever removed. We refer to all expressions that appear below the root expression as sub-expressions (including the root expression). By not ever removing any expressions, we keep all non-sub-expressions indefinitely. This potentially requires a lot of memory, but avoids explicit reference counting or garbage collection.

exprs_inv: HashMap<u64, Vec<usize>>

Maps expressions to their identifiers.

Serves as a fast inverse lookup for the unique identifier of a given sub-expression. To simplify ownership, we implement this lookup by mapping from the hash of a sub-expression to several identifiers. By structural sharing, the identifier for a sub-expression should be unique, but we still need a Vec for two reasons: First, there might be hash collisions, which we address by checking true equality when reading Formula::exprs_inv. Second, while sub-expressions have a unique identifier, there might be distinct, orphaned expressions that are equal to a given sub-expression. For example, such a situation arises when Formula::set_expr modifies a sub-expression and the resulting expression is equal (but not identical) to an existing sub-expression. As an expression cannot easily update its own identifier in the whole syntax tree, Formula::set_expr considers the first identifier in Formula::exprs_inv to be the canonical one (modulo hash collisions). Whenever Formula::set_expr encounters the concerned expression later, it then adapts its identifier to the canonical one. By this design, Formula::exprs_inv indeed maps any sub-expression (precisely: its hash) to its unique identifier (precisely: the first identifier whose expression is equal to the given sub-expression). This information can be used to enforce structural sharing by calling Formula::canon_visitor.

vars: Vec<Var<'a>>

Stores all variables in this formula.

Conceptually, this is analogous to Formula::exprs. However, there is no distinction analogous to sub-expressions and expressions, as variables need not be removed. Consequently, feature-model slicing (variable forgetting) is currently not supported. Another difference to Formula::exprs is that named variables are not owned by this formula. Thus, we can borrow references to variable names from the parsed string and avoid cloning them.

vars_inv: HashMap<Var<'a>, i32>

Maps variables to their identifiers.

Conceptually, this is analogous to Formula::exprs_inv. However, the inverse lookup of variables is less complex: First, this formula does not own the variable names, which avoids the hash collisions discussed for Formula::exprs_inv. Second, variables and their identifiers are never mutated after creation, so no additional Vec is needed.

root_id: usize

Specifies the root expression of this formula.

Serves as an index into Formula::exprs. The corresponding expression is the root of this formula’s syntax tree and thus the starting point for most algorithms. We consider all expressions below this expression (including itself) to be sub-expressions. There might be other (non-sub-)expressions that are currently not relevant to this formula.

var_aux_id: u32

Specifies the identifier of the most recently added auxiliary variable.

Ensures that new auxiliary variables (created with Var::Aux) are uniquely identified in the context of this formula.

new_exprs: Option<Vec<usize>>

Stores new expressions created by an algorithm but not yet included in the syntax tree.

Used by Formula::to_cnf_tseitin for holding on to definitional expressions.

Implementations

impl<'a> Formula<'a>

Algorithms for constructing, mutating, and analyzing formulas.

pub(crate) fn new() -> Self

Creates a new, empty formula.

The created formula is initially invalid (see Formula::assert_valid). The auxiliary variable with number 0 has no meaningful sign and can therefore not be used. This simplifies the representation of literals in crate::core::clauses::Clauses, which can be negative.

pub(crate) fn assert_valid(self) -> Self

Panics if this formula is invalid.

A formula is valid if it has at least one variable (added with Formula::var_expr) and a root expression (set with Formula::set_root_expr). In addition, structural sharing must not be violated (see Formula::canon_visitor). All assertions are optional and therefore not included in cargo build --release.

fn add_expr(&mut self, expr: Expr) -> usize

Adds a new expression to this formula, returning its new identifier.

Appends the given expression to Formula::exprs and enables its lookup via Formula::exprs_inv. Requires that no expression equal to the given expression is already in this formula. Thus, the created identifier will become the expression’s canonical identifier (see Formula::exprs_inv).

fn get_expr(&self, expr: &Expr) -> Option<usize>

Looks ups the identifier for an expression of this formula.

The canonical identifier for a given expression is the first one that is associated with its hash and whose expression is also equal to the given expression (see Formula::exprs_inv).

pub(crate) fn expr(&mut self, expr: Expr) -> usize

Adds or looks up an expression of this formula, returning its identifier.

This is the preferred way to obtain an expression’s identifier, as it ensures structural sharing. That is, the expression is only added to this formula if it does not already exist. Before we add the expression, we simplify it, which is a cheap operation (in contrast to Formula::flatten_expr).

fn add_var(&mut self, var: Var<'a>) -> i32

Adds a new variable to this formula, returning its identifier.

Works analogously to Formula::add_expr (see Formula::vars_inv).

fn add_var_named(&mut self, name: &'a str) -> i32

Adds a new named variable to this formula, returning its identifier.

pub(crate) fn add_var_aux(&mut self) -> i32

Adds a new auxiliary variable to this formula, returning its identifier.

fn get_var_named(&mut self, name: &str) -> Option<i32>

Looks ups the identifier of a named variable in this formula.

Works analogously to Formula::get_expr (see Formula::vars_inv).

pub(crate) fn var_expr(&mut self, var: &'a str) -> usize

Adds or looks up a named variable of this formula, returning its Var expression’s identifier.

This is the preferred way to obtain a Var expression’s identifier (see Formula::expr).

pub(crate) fn add_var_aux_expr(&mut self) -> usize

Adds a new auxiliary variable to this formula, returning its Var expression’s identifier.

pub(crate) fn get_root_expr(&self) -> usize

Returns the root expression of this formula.

pub(crate) fn set_root_expr(&mut self, root_id: usize)

Sets the root expression of this formula.

For a formula to be valid, the root expression has to be set at least once. It may also be updated subsequently to focus on other expressions of the formula or build more complex expressions.

fn simp_expr(&mut self, expr: &mut Expr)

Simplifies an expression in this formula to an equivalent one.

First, we sort the expression’s children, thus equality is up to commutativity. Second, we remove duplicate children of the expressions, thus equality is up to idempotency. Third, we identify unary expressions with their operands (i.e., And(x) is simplified to x). Fourth, we remove double negations (i.e., Not(Not(x)) is simplified to x). Fifth, we remove obvious tautologies and contradictions (i.e., And(a, Not(a)) is simplified to Or()). Because we clone expressions, this function may violate structural sharing (see Formula::canon_visitor). As this is a cheap and useful operation to make the formula smaller, we already call it in the parsing stage.

fn flatten_expr(&mut self, expr: &mut Expr)

Flattens children of an expression into their parent.

Analogously to Formula::simp_expr, this performs a simplification of an expression. However, this may create new expressions and is therefore more expensive and not called in the parsing stage. This is useful to call during a postorder syntax tree traversal to ensure canonical form (see Formula::canon_visitor).

fn inval_expr(&mut self, id: usize)

Invalidates an expression after it was mutated.

Does so by updating its mapping in Formula::exprs_inv. One of two cases applies, which can both be handled in the same way: Either the new expression has never been added before, so structural sharing was not violated. Thus, we can just append the expression’s identifier as the new canonical identifier for the expression. In the second case, the expression already exists and already has a canonical identifier. Still, we can append the identifier anyway, as only the first identifier will be considered. In terms of correctness, appending the identifier suffices, although we may optimize by cleaning up Formula::exprs_inv.

fn set_expr(&mut self, id: usize, expr: Expr)

Mutates an expression in this formula.

This function replaces the expression for a given identifier with a new given expression. It has no effect on leaves in the syntax tree (i.e., variables). We must take several precautions to preserve structural sharing, as we perform an in-place mutation. While this function may temporarily violate structural sharing when called for a given expression, it also makes up for (i.e., “fixes”) said violation when called for every parent of said expression afterwards (see Formula::canon_visitor). To do so, the function performs three steps: First, every new child expression is checked for potential duplicates with existing expressions, which we resolve using the canonical identifier obtained with Formula::get_expr. Second, we replace the old expression with the new expression. Third, as we might have changed the hash of the expression, we must invalidate it with Formula::inval_expr. Because this function cleans up violations of children, it must be called after, not before children have been mutated. Thus, it does not preserve structural sharing when used in Formula::preorder_rev, only in Formula::postorder_rev. Besides guaranteeing structural sharing, we perform flattening and simplification on the expression, which usually produces smaller formulas.

fn reset_root_expr(&mut self)

Resets the root expression, if necessary.

If the root expression is mutated with Formula::set_expr, structural sharing might be violated. Because Formula::set_expr can only address this issue for children, we need not explicitly address the only expression that is not a child itself - the root expression.

fn negate_exprs(&mut self, ids: Vec<usize>) -> Vec<usize>

Returns expressions that negate the given expressions.

The returned expression identifiers are either created or looked up (see Formula::expr).

fn format_expr(&self, id: usize, f: &mut Formatter<'_>) -> Result

Writes an expression of this formula to a formatter.

Used by fmt::Display to print (parts of) a formula. Implements a recursive preorder traversal. For an iterative reversed preorder traversal, see Formula::preorder_rev.

fn preorder_rev( &mut self, first_id: usize, visitor: impl FnMut(&mut Self, usize) )

Visits all sub-expressions of this formula using a reverse preorder traversal.

We assume that the given visitor is idempotent (in terms of formula mutation) and only performs mutation with the designated methods, such as Formula::var_expr, Formula::expr and Formula::set_expr. The visitor is called at most once per unique sub-expression: It will not be called several times on the same sub-expression as long (as it does not violate structural sharing). However, we can also not guarantee it to be called on all sub-expressions - as it might change the very set of sub-expressions. For improved performance, the traversal is reversed, so children are traversed right-to-left.

fn postorder_rev( &mut self, first_id: usize, visitor: impl FnMut(&mut Self, usize) )

Visits all sub-expressions of this formula using a reverse postorder traversal.

Conceptually, this is similar to Formula::preorder_rev, but sub-expressions are visited bottom-up instead of top-down. Also, this traversal can be used to ensure structural sharing if the visitor is correctly implemented (see Formula::canon_visitor).

fn prepostorder_rev( &mut self, first_id: usize, pre_visitor: impl FnMut(&mut Self, usize), post_visitor: impl FnMut(&mut Self, usize) )

Visits all sub-expressions of this formula using a combined reverse pre- and postorder traversal.

Can be used to efficiently interleave a preorder and postorder visitor. Note that each interior expression is visited twice (with the pre- and then the postorder visitor). However, the leaves (i.e., Var expressions) are only visited once (with the postorder visitor).

pub(crate) fn sub_exprs(&mut self) -> Vec<usize>

Returns the identifiers of all sub-expressions of this formula.

If in canonical form, each identifier is guaranteed to appear only once.

fn assert_canon(&mut self)

Panics if structural sharing is violated in this formula.

That is, we assert that every sub-expression’s identifier is indeed the canonical one. Does not currently check for commutativity, idempotency, or unary expressions.

fn canon_visitor(&mut self, id: usize)

Transforms this formula into canonical form.

In canonical form, several useful guarantees hold: First, no sub-expression occurs twice in the syntax tree with different identifiers (structural sharing). Second, equality of sub-expressions is up to commutativity, idempotency, and unary expressions. Third, no And expression is below an And expression (and analogously for Or). Fourth, no Not expression is below a Not expression. To ensure these guarantees, this visitor must be called in a postorder traversal, preorder does not work.

fn nnf_visitor(&mut self, id: usize)

Transforms this formula into negation normal form by applying De Morgan’s laws.

We do this by traversing the formula top-down, eanwhile, we push negations towards the leaves (i.e., Var expressions). Double negations cannot be encountered, as they have already been removed by Formula::simp_expr.

fn cnf_dist_visitor(&mut self, id: usize)

Transforms this formula into canonical conjunctive normal form by applying distributivity laws.

We do this by traversing the formula bottom-up and pushing Or expressions below And expressions via multiplication. This algorithm has exponential worst-case complexity, but ensures logical equivalence to the original formula. Assumes that this formula is already in negation normal form.

fn def_and(&mut self, var_expr_id: usize, ids: &[usize]) -> (i32, Vec<usize>)

Defines an And expression with a new auxiliary variable.

That is, we create a new auxiliary variable and clauses that let it imply all conjuncts and let it be implied by the conjunction. As an optimization, we do not create a Var expression for the new variable, as we are replacing an existing expression.

fn def_or(&mut self, var_expr_id: usize, ids: &[usize]) -> (i32, Vec<usize>)

Defines an Or expression with a new auxiliary variable.

That is, we create a new auxiliary variable and clauses that let it imply the disjunction and let it be implied by all disjuncts. Works analogously to Formula::def_and.

fn cnf_tseitin_visitor(&mut self, id: usize)

pub(crate) fn to_nnf(self) -> Self

Transforms this formula into canonical negation normal form by applying De Morgan’s laws.

pub(crate) fn to_cnf_dist(self) -> Self

Transforms this formula into canonical conjunctive normal form by applying distributivity laws.

pub(crate) fn to_cnf_tseitin(self) -> Self

Transforms this formula into canonical conjunctive normal form by TODO