clausy::util

Module io

Utilities for handling UVL, XML, and other files.

Functions

name_from_io ğŸ”’

Retrieves a name from a given form compatible with FeatureIDE.

name_to_io ğŸ”’

Transforms a given name into a form compatible with FeatureIDE.

to_uvl_string ğŸ”’

Expresses a clause representation as a string of UVL features and constraints.

to_xml_string ğŸ”’

Expresses a clause representation as a string of XML constraints.

uvl_add_vars ğŸ”’

Appends variables as abstract features to an existing UVL feature hierarchy.

uvl_append ğŸ”’

Given a UVL file with a feature hierarchy and given features or constraints, creates a merged UVL file.

uvl_file_add_vars ğŸ”’

Appends variables by their identifiers to an existing UVL file.

uvl_remove_constraints ğŸ”’

Given a UVL file, only returns its feature hierarchy, omitting its constraints.

write_constraints ğŸ”’

Writes constraints to a file.

write_uvl_and_xml ğŸ”’

Writes a UVL and XML file with a given prefix.

write_vars ğŸ”’

Writes variables to a file.

xml_with_constraints ğŸ”’

Given a XML file with a feature hierarchy and given constraints, creates a merged XML file.