Struct clausy::core::arena::Arena

pub(crate) struct Arena {
    pub(super) vars: Vec<Var>,
    vars_inv: HashMap<Var, i32>,
    pub(super) exprs: Vec<Expr>,
    exprs_inv: HashMap<u64, Vec<usize>>,
    var_aux_id: u32,
    pub(super) new_vars: Option<Vec<i32>>,
    pub(super) new_exprs: Option<Vec<usize>>,
}

An arena of variables and expressions.

We aim to represent any Formula by storing its syntax tree; that is, each unique sub-expression that appears in it. To allow reuse of expressions and variables across similar formulas, we store them in a shared arena referenced by each Formula. In a canonical Formula, sub-expressions are uniquely stored, so no sub-expression appears twice with distinct identifiers (structural sharing). This allows for concise representation and facilitates some algorithms (e.g., Formula::to_cnf_tseitin). However, it also comes with the downside that each sub-expression has potentially many parents. Thus, owners of sub-expressions are not easily trackable (see Arena::exprs on garbage collection). Consequently, we cannot access any parents when mutating sub-expressions, only children. Due to structural sharing, we effectively treat the syntax tree as a directed acyclic graph. We represent this graph as an adjacency list stored in Arena::exprs. Note that, for performance reasons, structural sharing is not guaranteed by all algorithms (including parsers) until calling Arena::canon_visitor.

Fields
vars: Vec<Var>

Stores all variables in this arena.

Conceptually, this is analogous to Arena::exprs. Accordingly, the set of variables referenced by a given Formula constitutes its sub-variables (see Formula::sub_var_ids). Variable-forgetting operations, such as feature-model slicing, are possible by adapting this set.

vars_inv: HashMap<Var, i32>

Maps variables to their identifiers.

Conceptually, this is analogous to Arena::exprs_inv. However, we simplify the inverse lookup of variables by cloning variables. This needs more memory, but (a) we do not expect many variables and (b) this avoids the hash collisions discussed for Arena::exprs_inv. Also, variables and their identifiers are never mutated after creation, so we need no additional Vec for updates.

exprs: Vec<Expr>

Stores all expressions in this arena.

Serves as a fast lookup for an expression, given its identifier. Expressions are stored in the order of their creation, so new expressions are appended with Vec::push. Also, while algorithms may update expressions in-place, no expression is ever removed. We refer to all expressions that appear below the root expression of a Formula as its sub-expressions (including the root expression). By not ever removing any expressions, we keep all non-sub-expressions indefinitely. This potentially requires a lot of memory, but avoids explicit reference counting or garbage collection.

exprs_inv: HashMap<u64, Vec<usize>>

Maps expressions to their identifiers.

Serves as a fast inverse lookup for the unique identifier of a given sub-expression. To simplify ownership, we implement this lookup by mapping from the hash of a sub-expression to several identifiers. By structural sharing, the identifier for a sub-expression should be unique, but we still need a Vec for two reasons: First, there might be hash collisions, which we address by checking true equality when reading Arena::exprs_inv. Second, while sub-expressions have a unique identifier, there might be distinct, orphaned expressions that are equal to a given sub-expression. For example, such a situation arises when Arena::set_expr modifies a sub-expression and the resulting expression is equal (but not identical) to an existing sub-expression. As an expression cannot easily update its own identifier in the whole syntax tree, Arena::set_expr considers the first identifier in Arena::exprs_inv to be the canonical one (modulo hash collisions). Whenever Arena::set_expr encounters the concerned expression later, it then adapts its identifier to the canonical one. By this design, Arena::exprs_inv indeed maps any sub-expression (precisely: its hash) to its unique identifier (precisely: the first identifier whose expression is equal to the given sub-expression). This information can be used to enforce structural sharing by calling Arena::canon_visitor.

var_aux_id: u32

Specifies the identifier of the most recently added auxiliary variable.

Ensures that new auxiliary variables (created with Var::Aux) are uniquely identified in the context of this arena.

new_vars: Option<Vec<i32>>

Temporarily stores new variables created that are not yet included in any Formula’s syntax tree.

Used by Arena::cnf_tseitin_visitor for holding on to auxiliary variables.

new_exprs: Option<Vec<usize>>

Temporarily stores new expressions that are not yet included in any Formula’s syntax tree.

Used by Arena::cnf_tseitin_visitor for holding on to definitional expressions.

Implementations

impl Arena

pub(crate) fn new() -> Self

Creates a new, empty arena.

This simplifies the representation of literals in crate::core::clauses::Clauses, which can be negative.

fn add_var(&mut self, var: Var) -> i32

Adds a new variable to this arena, returning its identifier.

Works analogously to Arena::add_expr (see Arena::vars_inv).

fn add_var_named(&mut self, name: String) -> i32

Adds a new named variable to this arena, returning its identifier.

pub(super) fn add_var_aux(&mut self) -> i32

Adds a new auxiliary variable to this arena, returning its identifier.

fn get_var_named(&mut self, name: String) -> Option<i32>

Looks ups the identifier of a named variable in this arena.

Works analogously to Arena::get_expr (see Arena::vars_inv).

pub(crate) fn var_expr(&mut self, var: String) -> usize

Adds or looks up a named variable in this arena, returning its Var expression’s identifier.

This is the preferred way to obtain a Var expression’s identifier (see Arena::expr).

pub(crate) fn var_expr_with_id(&mut self, var: String) -> (usize, i32)

Adds or looks up a named variable in this arena, returning its Var expression’s and Var’s identifier.

pub(crate) fn add_var_aux_expr(&mut self) -> (i32, usize)

Adds a new auxiliary variable to this arena, returning its identifier and its Var expression’s identifier.

pub(crate) fn vars( &self, predicate: impl Fn(i32, &Var) -> bool ) -> Vec<(i32, Var)>

Returns all variables and their identifiers in this arena that match a given predicate.

fn add_expr(&mut self, expr: Expr) -> usize

Adds a new expression to this arena, returning its new identifier.

Appends the given expression to Arena::exprs and enables its lookup via Arena::exprs_inv. Requires that no expression equal to the given expression is already in this arena. Thus, the created identifier will become the expression’s canonical identifier (see Arena::exprs_inv).

pub(super) fn get_expr(&self, expr: &Expr) -> Option<usize>

Looks ups the identifier for an expression in this arena.

The canonical identifier for a given expression is the first one that is associated with its hash and whose expression is also equal to the given expression (see Arena::exprs_inv).

pub(crate) fn expr(&mut self, expr: Expr) -> usize

Adds or looks up an expression in this arena, returning its identifier.

This is the preferred way to obtain an expression’s identifier, as it ensures structural sharing. That is, the expression is only added to this arena if it does not already exist. Before we add the expression, we simplify it, which is a cheap operation (in contrast to Arena::flatten_expr).

fn simp_expr(&mut self, expr: &mut Expr)

Simplifies an expression in this arena to an equivalent one.

First, we sort the expression’s children, thus equality is up to commutativity. Second, we remove duplicate children of the expressions, thus equality is up to idempotency. Third, we identify unary expressions with their operands (i.e., And(x) is simplified to x). Fourth, we remove double negations (i.e., Not(Not(x)) is simplified to x). Fifth, we remove obvious tautologies and contradictions (i.e., And(a, Not(a)) is simplified to Or()). Because we clone expressions, this function may violate structural sharing (see Arena::canon_visitor). As this is a cheap and useful operation to make the arena smaller, we already call it in the parsing stage.

fn flatten_expr(&mut self, expr: &mut Expr)

Flattens children of an expression into their parent.

Analogously to Arena::simp_expr, this performs a simplification of an expression. However, this may create new expressions and is therefore more expensive, thus not called in the parsing stage. This is useful to call during a postorder syntax tree traversal to ensure canonical form (see Arena::canon_visitor).

fn inval_expr(&mut self, id: usize)

Invalidates an expression after it was mutated.

Does so by updating its mapping in Arena::exprs_inv. One of two cases applies, which can both be handled in the same way: Either the new expression has never been added before, so structural sharing was not violated. Thus, we can just append the expression’s identifier as the new canonical identifier for the expression. In the second case, the expression already exists and already has a canonical identifier. Still, we can append the identifier anyway, as only the first identifier will be considered. In terms of correctness, appending the identifier suffices, although we may optimize in the future by cleaning up Arena::exprs_inv.

fn set_expr(&mut self, id: usize, expr: Expr)

Mutates an expression in this arena.

This function replaces the expression for a given identifier with a new given expression. It has no effect on leaves in the syntax tree (i.e., variables). We must take several precautions to preserve structural sharing, as we perform an in-place mutation. While this function may temporarily violate structural sharing when called for a given expression, it also makes up for (i.e., “fixes”) said violation when called for every parent of said expression afterwards (see Arena::canon_visitor). To do so, the function performs three steps: First, every new child expression is checked for potential duplicates with existing expressions, which we resolve using the canonical identifier obtained with Arena::get_expr. Second, we replace the old expression with the new expression. Third, as we might have changed the hash of the expression, we must invalidate it with Arena::inval_expr. Because this function cleans up violations of children, it must be called after, not before children have been mutated. Thus, it does not preserve structural sharing when used in Arena::preorder_rev, only in Arena::postorder_rev. Besides guaranteeing structural sharing, we perform flattening and simplification on the expression, which usually produces smaller formulas.

fn negate_exprs(&mut self, ids: Vec<usize>) -> Vec<usize>

Returns expressions that negate the given expressions.

The returned expression identifiers are either created or looked up (see Arena::expr).

pub(super) fn format_expr(&self, id: usize, f: &mut Formatter<'_>) -> Result

Writes an expression in this arena to a formatter.

Used by fmt::Display to print (parts of) a Formula. Implements a recursive preorder traversal. For an iterative reversed preorder traversal, see Arena::preorder_rev.

pub(crate) fn as_formula<'a>(&'a self, root_id: usize) -> Formula

Returns a formula with the given root expression.

The created formula references all variables of this arena, use Formula::new for more fine-grained sub-variables.

pub(super) fn preorder_rev( &mut self, root_id: &mut usize, visitor: impl FnMut(&mut Self, usize) )

Visits all sub-expressions of a given root expression using a reverse preorder traversal.

We assume that the given visitor only performs mutation with the designated methods, such as Arena::var_expr, Arena::expr and Arena::set_expr. The visitor is called at most once per unique sub-expression: It will not be called several times on the same sub-expression (if the formula is in canonical form). However, we can also not guarantee it to be called on all sub-expressions - as it might change the very set of sub-expressions. For improved performance, the traversal is reversed, so children are traversed right-to-left.

pub(super) fn postorder_rev( &mut self, root_id: &mut usize, visitor: impl FnMut(&mut Self, usize) )

Visits all sub-expressions of a given root expression using a reverse postorder traversal.

Conceptually, this is similar to Arena::preorder_rev, but sub-expressions are visited bottom-up instead of top-down. Also, this traversal can be used to ensure structural sharing if the visitor is correctly implemented (see Arena::canon_visitor).

pub(super) fn prepostorder_rev( &mut self, root_id: &mut usize, pre_visitor: impl FnMut(&mut Self, usize), post_visitor: impl FnMut(&mut Self, usize) )

Visits all sub-expressions of a given root expression using a combined reverse pre- and postorder traversal.

Can be used to efficiently interleave a preorder and postorder visitor. Note that each interior expression is visited twice (with the pre- and then the postorder visitor). However, the leaves (i.e., Var expressions) are only visited once (with the postorder visitor).

pub(super) fn canon_visitor(&mut self, id: usize)

Transforms an expression into canonical form (see Formula::to_canon).

pub(super) fn nnf_visitor(&mut self, id: usize)

Transforms an expression into negation normal form by applying De Morgan’s laws (see Formula::to_nnf).

pub(super) fn cnf_dist_visitor(&mut self, id: usize)

Transforms an expression into canonical conjunctive normal form by applying distributivity laws (see Formula::to_cnf_dist).

fn def_and(&mut self, var_expr_id: usize, ids: &[usize]) -> i32

Defines an And expression with a new auxiliary variable.

That is, we create a new auxiliary variable and clauses that let it imply all conjuncts and let it be implied by the conjunction. As an optimization, we do not create a Var expression for the new variable, as we are replacing an existing expression. We add the new variable to Arena::new_vars and the clauses defining the new variable to Arena::new_exprs.

fn def_or(&mut self, var_expr_id: usize, ids: &[usize]) -> i32

Defines an Or expression with a new auxiliary variable.

That is, we create a new auxiliary variable and clauses that let it imply the disjunction and let it be implied by all disjuncts. Works analogously to Arena::def_and.

pub(super) fn cnf_tseitin_visitor(&mut self, id: usize)

Transforms an expression into canonical conjunctive normal form by introducing auxiliary variables (see Formula::to_cnf_tseitin).