Struct clausy::core::formula::Formula

pub(crate) struct Formula {
    pub(crate) sub_var_ids: HashSet<i32>,
    pub(crate) root_id: usize,
    pub(crate) file: Option<File>,
}

A feature-model formula.

A Formula is a view onto part of an Arena, which contains the actual implementation of most algorithms on formulas. As such, a formula is given by the set of variables (defining its universe of solutions) as well as the root expression of its syntax tree (constraining said variables). We must store the variables to ensure correct results for certain feature-model analyses (e.g., model counting and slicing). A formula is only a view and always implicitly tied to an Arena.

Fields
sub_var_ids: HashSet<i32>

Specifies the sub-variables of this formula.

Each identifiers serves as an index into Arena::vars. All variables in the syntax tree of this formula must occur in this set, but it may contain more (unconstrained) variables.

root_id: usize

Specifies the root expression of this formula.

Serves as an index into Arena::exprs. The corresponding expression is the root of this formula’s syntax tree and thus the starting point for most algorithms. We consider all expressions below this expression (including itself) to be sub-expressions. There might be other (non-sub-)expressions that are currently not relevant to this formula.

file: Option<File>

The file this formula was originally parsed from, if any.

Implementations

impl Formula

pub(crate) fn new( sub_var_ids: HashSet<i32>, root_id: usize, file: Option<File> ) -> Self

Creates a new formula.

The sub-variable and root expression identifiers must be valid in the context of some given Arena.

pub(crate) fn as_ref<'a>(&'a self, arena: &'a Arena) -> FormulaRef<'_>

Returns a shared reference to this formula in the context of its arena.

pub(crate) fn to_clauses(&self, arena: &Arena) -> Clauses

Returns the clause representation of this formula.

pub(super) fn reset_root_expr(root_id: &mut usize, arena: &Arena)

Resets the root expression of this formula, if necessary.

If the root expression is mutated with Arena::set_expr, structural sharing might be violated. Because Arena::set_expr can only address this issue for children, we need to explicitly address the only expression that is not a child itself - the root expression.

pub(crate) fn force_foreign_vars( &self, top: bool, exclude_vars: &HashSet<i32>, arena: &mut Arena ) -> Formula

Returns a formula that forces all variables only occurring in the given arena to true or false.

Does not modify this formula.

pub(crate) fn sub_vars(&self, arena: &Arena) -> Vec<(i32, Var)>

Returns all sub-variables of this formula and their identifiers.

pub(crate) fn common_vars(&self, other: &Formula) -> HashSet<i32>

Returns all sub-variable identifiers of this formula also in another formula.

pub(crate) fn all_vars(&self, other: &Formula) -> HashSet<i32>

Returns all sub-variable identifiers of this formula or in another formula.

pub(crate) fn except_vars(&self, other: &Formula) -> HashSet<i32>

Returns all sub-variable identifiers of this formula not in another formula.

pub(crate) fn sub_exprs(&mut self, arena: &mut Arena) -> Vec<usize>

Returns the identifiers of all sub-expressions of this formula.

If in canonical form, each identifier is guaranteed to appear only once.

pub(crate) fn assert_canon(&mut self, arena: &mut Arena)

Panics if structural sharing is violated in this formula.

That is, we assert that every sub-expression’s identifier is indeed the canonical one. Does not currently check for other properties of canonicity (see Formula::to_canon).

pub(crate) fn to_canon(&mut self, arena: &mut Arena)

Transforms this formula into canonical form (see Arena::canon_visitor).

The resulting formula is logically equivalent to the original formula. This function is useful when an algorithm assumes or profits from canonical form, or for simplifying a formula after parsing. In canonical form, several useful guarantees hold: First, no sub-expression occurs twice in the syntax tree with different identifiers (structural sharing). Second, equality of sub-expressions is up to commutativity, idempotency, and unary expressions, and those expressions are simplified. Third, no And expression is below an And expression (and analogously for Or). Fourth, no Not expression is below a Not expression. To ensure these guarantees, this visitor must be called in a postorder traversal, preorder does not work.

pub(crate) fn to_nnf(&mut self, arena: &mut Arena)

Transforms this formula into canonical negation normal form by applying De Morgan’s laws (see Arena::nnf_visitor).

The resulting formula is logically equivalent to the original formula. We do this by traversing the formula top-down, meanwhile, we push negations towards the leaves (i.e., Var expressions). Double negations cannot be encountered, as they have already been removed by Arena::simp_expr.

pub(crate) fn to_cnf_dist(&mut self, arena: &mut Arena)

Transforms this formula into canonical conjunctive normal form by applying distributivity laws (see Arena::cnf_dist_visitor).

The resulting formula is logically equivalent to the original formula. We do this by traversing the formula bottom-up and pushing Or expressions below And expressions via multiplication. This algorithm has exponential worst-case complexity, but ensures logical equivalence to the original formula.

pub(crate) fn to_cnf_tseitin(&mut self, assume_root: bool, arena: &mut Arena)

Transforms this formula into canonical conjunctive normal form by introducing auxiliary variables (see Arena::cnf_tseitin_visitor).

The resulting formula is equivalent to the original formula projected onto its named variables (i.e., satisfiability and model count are preserved). If this formula is in canonical form (see Formula::to_canon), we introduce exactly one auxiliary variable per (complex) sub-expression. Thus, every sub-expression will be “abbreviated” with an auxiliary variable, including the root expression, which facilitates algebraic operations. Also, no sub-expression will be abbreviated twice, so the number of auxiliary variables is equal to the number of sub-expressions. If this formula is not in canonical form, more auxiliary variables might be introduced. Note that we only abbreviate complex sub-expressions (i.e., And and Or expressions), as Var and Not expressions do not profit from abbrevation.

pub(crate) fn remove_constraints( &self, ids: &HashSet<i32>, arena: &mut Arena ) -> Formula

Returns a formula that only contains constraints of this formula that do not contain any given variable.

Assumes that this formula is in proto-CNF; that is, it is a conjunction of constraints. Does not modify this formula.

pub(crate) fn assert_proto_cnf(&self, arena: &Arena)

Asserts that this formula is in proto-CNF; that is, it is a non-empty conjunction of constraints.

pub(crate) fn diff_vars( &self, other: &Formula ) -> (HashSet<i32>, HashSet<i32>, HashSet<i32>)

Returns the identifiers of all variables common and unique to this and a given formula.

pub(crate) fn diff_constraints( &self, other: &Formula, arena: &mut Arena ) -> (HashSet<usize>, HashSet<usize>, HashSet<usize>)

Returns the identifiers of all constraints common and unique to this and a given formula.

Assumes that this formula is in proto-CNF; that is, it is a conjunction of constraints.

pub(crate) fn assume(&mut self, id: usize, arena: &mut Arena) -> Formula

Returns a formula that assumes an additional constraint.

pub(crate) fn and_expr(&self, other: &Formula, arena: &mut Arena) -> usize

Returns an expression that encodes the common solutions of this and another formula.

pub(crate) fn and(&self, other: &Formula, arena: &mut Arena) -> Formula

Returns a formula that encodes the common solutions of this and another formula.

Does not modify this formula.

pub(crate) fn implies_expr(&self, other: &Formula, arena: &mut Arena) -> usize

Returns an expression that encodes whether this formula implies another formula.

Also encodes solutions gone in the other formula, if any.

pub(crate) fn implies(&self, other: &Formula, arena: &mut Arena) -> Formula

Returns a formula that encodes whether this formula implies another formula.

Does not modify this formula.

pub(crate) fn analyze( &self, arena: &Arena, use_tseitin: bool, count: bool, serialize: bool ) -> (BigInt, Option<String>, Option<String>)

Returns the number of solutions of this formula or serializes it.

Optionally uses a Tseitin transformation into CNF. Does not modify this formula or the given arena.

pub(crate) fn count_inc( &self, b: &Formula, left_model_count: Option<&str>, arena: &mut Arena ) -> String

Returns a mathematical term that, given the number of solutions of this formula, calculates the number of solutions of another formula.

pub(crate) fn diff( &self, b: &Formula, left_diff_kind: DiffKind, right_diff_kind: DiffKind, prefix: Option<&str>, arena: &mut Arena )

Prints or serializes a description of the difference between this formula and another.

Assumes that common variables are considered equal (e.g., equal features have equal names), that the input formulas contains no auxiliary variables, and that the input formulas are in proto-CNF.