Struct clausy::formula::Formula

pub struct Formula<'a> {
    pub(crate) exprs: Vec<Expr>,
    exprs_inv: HashMap<u64, Vec<usize>>,
    pub(crate) vars: Vec<&'a str>,
    vars_inv: HashMap<&'a str, i32>,
    aux_root_id: usize,
}

A feature-model formula.

We represent a formula by storing its syntax tree; that is, each unique sub-expression that appears in it. As an invariant, sub-expressions are uniquely stored, so no sub-expression can appear twice with distinct identifiers (structural sharing). This invariant allows for concise representation and facilitates some algorithms (e.g., Tseitin transformation). However, it also comes with the downside that each sub-expression has potentially many parents. Thus, owners of sub-expressions are not easily trackable (see Formula::exprs on garbage collection). Consequently, all algorithms must be implemented in a way that only mutates the children of an expression, not their parent(s).

Fields
exprs: Vec<Expr>

Stores all expressions in this formula.

Serves as a fast lookup for an expression, given its identifier. Expressions are stored in the order of their creation, so new expressions are appended with Vec::push. Also, while some algorithms may update expressions in-place, no expression is ever removed. We refer to all expressions that appear below the auxiliary root expression as sub-expressions. By not ever removing any expressions, we keep all non-sub-expressions indefinitely. This potentially requires a lot of memory, but avoids explicit reference counting or garbage collection.

exprs_inv: HashMap<u64, Vec<usize>>

Maps expressions to their identifiers.

Serves as a fast inverse lookup for the unique identifier of a given sub-expression. To simplify ownership, we implement this lookup by mapping from the hash of a sub-expression to several identifiers. By structural sharing, the identifier for a sub-expression should be unique, but we still need a Vec for two reasons: First, there might be hash collisions, which we address by checking true equality when reading Formula::exprs_inv. Second, while sub-expressions have a unique identifier, there might be distinct, orphaned expressions that are equal to a given sub-expression. For example, such a situation arises when Formula::set_child_exprs modifies its expression’s children and the resulting expression is equal (but not identical) to an existing sub-expression. As an expression cannot easily change its own identifier (similar to how a variable cannot change its own type), Formula::set_child_exprs considers the first identifier in Formula::exprs_inv to be the canonical one (modulo hash collisions). Whenever Formula::set_child_exprs encounters the concerned expression, it then adapts its identifier to the canonical one. By this design, Formula::exprs_inv indeed maps any sub-expression (precisely: its hash) to its unique identifier (precisely: the first identifier whose expression is equal to the given sub-expression). Any algorithms that mutate this formula should take this into account to preserve structural sharing as an invariant.

vars: Vec<&'a str>

Stores all variables in this formula.

Conceptually, this is analogous to Formula::exprs. However, there is no distinction analogous to sub-expressions and expressions, as variables need not be removed. Consequently, feature-model slicing (variable forgetting) is currently not supported. Another difference to Formula::exprs is that variables (which are simply names) are not owned by this formula. Thus, we can borrow references to variable names from the parsed string and avoid cloning them.

vars_inv: HashMap<&'a str, i32>

Maps variables to their identifiers.

Conceptually, this is analogous to Formula::exprs_inv. However, the inverse lookup of variables is simpler: First, this formula does not own the variables, which avoids the hash collisions discussed for Formula::exprs_inv. Second, variables and their identifiers are never mutated after creation, so no additional Vec is needed.

aux_root_id: usize

Specifies the auxiliary root expression of this formula.

Serves as an index into Formula::exprs. The corresponding expression is the auxiliary root of this formula’s syntax tree and thus the starting point for most algorithms. We consider all expressions below this expression (including itself) to be sub-expressions. There might be other (non-sub-)expressions that are currently not relevant to this formula. Note that Formula::get_root_expr and Formula::set_root_expr do not store the user-supplied root expression here, but an auxiliary And expression that has the root expression as its single child. This allows algorithms to freely mutate the root expression if necessary (see Formula on mutating children).

Implementations

impl<'a> Formula<'a>

Algorithms for constructing, mutating, and analyzing formulas.

pub(crate) fn new() -> Self

Creates a new, empty formula.

The created formula is initially invalid (see Formula::assert_valid). The variable with empty name and identifier 0 has no meaningful sign and can therefore not be used. This simplifies the representation of literals in crate::cnf::CNF.

fn assert_valid(&self)

Panics if this formula is invalid.

A formula is valid if it has at least one variable (added with Formula::var) and a root expression (set with Formula::set_root_expr).

fn hash_expr(expr: &Expr) -> u64

Computes the hash of an expression.

Used to look up an expression’s identifier in Formula::exprs_inv.

fn add_expr(&mut self, expr: Expr) -> usize

Adds a new expression to this formula, returning its new identifier.

Appends the given expression to Formula::exprs and enables its lookup via Formula::exprs_inv. Requires that no expression equal to the given expression is already in this formula. Thus, the created identifier will become the expression’s canonical identifier (see Formula::exprs_inv).

fn get_expr(&self, expr: &Expr) -> Option<usize>

Looks ups the identifier for an expression of this formula.

The canonical identifier for a given expression is the first one that is associated with its hash and whose expression is also equal to the given expression (see Formula::exprs_inv).

pub(crate) fn expr(&mut self, expr: Expr) -> usize

Adds or looks up an expression of this formula, returning its identifier.

This is the preferred way to obtain an expression’s identifier, as it ensures structural sharing. That is, the expression is only added to this formula if it does not already exist.

fn add_var(&mut self, var: &'a str) -> usize

Adds a new variable to this formula, returning the identifier of its Var expression.

Works analogously to Formula::add_expr (see Formula::vars_inv). However, it does not return the variable’s identifier, but its Var expression’s identifier. This is usually more convenient.

fn get_var(&mut self, var: &str) -> Option<usize>

Looks ups the identifier for the Var expression of a variable in this formula.

Works analogously to Formula::get_expr (see Formula::vars_inv). As for Formula::add_var, it is usually more convenient to return the Var expression’s identifier.

pub(crate) fn var(&mut self, var: &'a str) -> usize

Adds or looks up a variable of this formula, returning its Var expression’s identifier.

This is the preferred way to obtain a Var expression’s identifier (see Formula::expr).

pub(crate) fn get_root_expr(&self) -> usize

Returns the root expression of this formula.

That is, we return the only child of the auxiliary root expression (see Formula::aux_root_id).

pub(crate) fn set_root_expr(&mut self, root_id: usize)

Sets the root expression of this formula.

That is, we update this formula’s auxiliary root expression with the given expression (see Formula::aux_root_id). For a formula to be valid, the root expression has to be set at least once. It may also be updated subsequently to focus on other expressions of the formula.

fn get_child_exprs<'b>(expr: &'b Expr) -> &'b [usize]

Returns the identifiers of the children of an expression.

We return nothing for Var expressions, which have no expression identifiers as children (only a variable identifier). As Var expressions are leaves of a formula’s syntax tree, this function is useful for traversing that tree.

fn set_child_exprs(&mut self, id: usize, ids: Vec<usize>)

Sets the children of an expression in this formula.

This function must take several precautions to preserve structural sharing, as it performs in-place mutations. While this function may temporarily violate structural sharing when called for a given expression, it also makes up for said violation when called for any parent of said expression. To do this, the function performs three steps: First, every new child expression is checked for potential duplicates with existing expressions, which we resolve using the canonical identifier obtained with Formula::get_expr. Second, we replace the old children with the new children. Third, as we might have changed the hash of the expression, we must update its mapping in Formula::exprs_inv. One of two cases applies, which can both be handled in the same way: Either the new expression has never been added before, so structural sharing was not violated. Thus, we can just push the expression’s identifier as the new canonical identifier for the expression. In the second case, the expression already exists and already has a canonical identifier. Still, we can push the identifier anyway, as only the first identifier will be considered.

fn reset_aux_root_expr(&mut self)

Resets the auxiliary root expression, if necessary.

If the auxiliary root expression is mutated with Formula::set_child_exprs, structural sharing might be violated. Because Formula::set_child_exprs can only address this issue for children, we need not explicitly address the only expression that is not a child itself - the auxiliary root expression.

fn negate_exprs(&mut self, ids: Vec<usize>) -> Vec<usize>

Returns expressions that negate the given expressions.

The returned expression identifiers are either created or looked up (see Formula::expr).

fn format_expr(&self, id: usize, f: &mut Formatter<'_>) -> Result

Writes an expression of this formula to a formatter.

Used by fmt::Display to print (parts of) a formula. Implements a recursive preorder traversal. For an iterative reversed preorder traversal, see Formula::preorder_rev.

fn is_non_aux_and(&self, id: usize) -> bool

TODO

fn splice_or(&self, clause_id: usize, new_clause: &mut Vec<usize>)

TODO

fn dedup(vec: Vec<usize>) -> Vec<usize>

TODO

fn preorder_rev(&mut self, visitor: fn(_: &mut Self, _: usize))

Visits all sub-expressions of this formula using a reverse preorder traversal.

To preserve structural sharing, we assume that the given visitor is idempotent and only performs mutation with the designated methods, such as Formula::var, Formula::expr and Formula::set_child_exprs. The visitor is called at most once per unique sub-expression: It will not be called several times on the same sub-expression - this leverages structural sharing. However, we can also not guarantee it to be called on all sub-expressions - as it might change the set of sub-expressions. For improved performance, the traversal is reversed, so children are traversed right-to-left.

fn postorder_rev(&mut self, visitor: fn(_: &mut Self, _: usize))

Visits all sub-expressions of this formula using a reverse postorder traversal.

Conceptually, this is similar to Formula::preorder_rev, but sub-expressions are visited bottom-up instead of top-down.

pub fn print_sub_exprs(&mut self)

Prints all sub-expression of this formula.

pub fn assert_shared(&mut self)

Panics if structural sharing is violated in this formula.

That is, we assert that every sub-expression’s identifier is indeed the canonical one.

pub fn to_nnf(self) -> Self

Transforms this formula into negation normal form by applying De Morgan’s laws.

We do this by traversing the formula top-down and pushing negations towards the leaves (i.e., Var expressions). TODO

pub fn to_cnf_dist(self) -> Self

Transforms this formula into conjunctive normal form by applying distributivity laws.

We do this by traversing the formula bottom-up and pushing Or expressions below And expressions via multiplication. This algorithm has exponential worst-case complexity, but ensures logical equivalence to the original formula. TODO TODO refactor code TODO is this idempotent?